Fix borking Bash on exit, remove `threadpool` dep
Prior to this commit, we borked some terminals running Bash on exit.
The thread dealing with user input is blocked waiting for it in a loop,
when the program quits, the thread gets culled by the OS and `getch`
proc doesn't get to restore the state of the stdin, so we do it manually
now. Thanks to @beef331 for the implementation of the exitproc.

`std/threadpool` wasn't really used, besides `spawn`, so it's removed.

Co-authored-by: Jason Beetham  <beefers331@gmail.com>